Move demo app to src/examples
