+++ tox.ini
@@ -48,15 +48,15 @@ commands =
 
     functests: pytest tests/functional/ -s
 
-    lint: pydocstyle --explain src dev
+    lint: pydocstyle --explain src
     lint: pydocstyle --config tests/.pydocstyle --explain tests
-    lint: pylint {posargs:src bin dev}
+    lint: pylint {posargs:src bin}
     lint: pylint --rcfile=tests/.pylintrc tests
 
-    format: black {posargs:src tests bin dev}
+    format: black {posargs:src tests bin}
     format: isort --atomic .
 
-    checkformatting: black --check {posargs:src tests bin dev}
+    checkformatting: black --check {posargs:src tests bin}
     checkformatting: isort --quiet --check-only .
 
     coverage: -coverage combine
